refactor: only warn once
