+++ tests/unit/services/test_project.py
@@ -19,10 +19,12 @@
 import itertools
 import pathlib
 from typing import Any
+from unittest.mock import call
 
 import pytest
 import pytest_mock
 from craft_application.errors import CraftValidationError
+from craft_cli.pytest_plugin import RecordingEmitter
 
 from snapcraft import const
 from snapcraft.application import APP_METADATA
@@ -30,11 +32,13 @@ from snapcraft.services.project import Project
 
 
 @pytest.fixture(autouse=True)
-def reset_ua_service_warning():
-    """Reset the one-shot ua-service warning flag between tests."""
+def reset_warnings():
+    """Reset the one-shot warning flags between tests."""
     Project._ua_service_warning = False
+    Project._license_spdx_warning = False
     yield
     Project._ua_service_warning = False
+    Project._license_spdx_warning = False
 
 
 @pytest.mark.parametrize(
@@ -158,3 +162,45 @@ class TestValidateUaServices:
         Project.validate_ua_services(project)
 
         emitter.assert_interactions(None)
+
+
+class TestValidateLicense:
+    def test_non_spdx_deprecation_warns_once(self, emitter: RecordingEmitter) -> None:
+        project = {"license": "maybe"}
+
+        Project.validate_license_spdx(project)
+        Project.validate_license_spdx(project)
+
+        emitter.assert_warning(
+            "Non-SPDX licenses are deprecated. Use SPDX license strings or 'proprietary' instead. For more information, see https://spdx.org/licenses/."
+        )
+        # assert it was only shown once
+        assert len(emitter.interactions) == 1
+
+    @pytest.mark.parametrize(
+        ("lic", "should_warn"),
+        [
+            ("MIT", False),
+            ("proprietary", False),
+            (None, False),
+            ("DemonicContract", True),
+        ],
+    )
+    def test_non_spdx_deprecation(
+        self,
+        lic: str | None,
+        should_warn: bool,
+        emitter: RecordingEmitter,
+    ) -> None:
+        project = {"license": lic}
+        Project.validate_license_spdx(project)
+
+        assert should_warn == (
+            call(
+                "warning",
+                "Non-SPDX licenses are deprecated. Use SPDX license strings or 'proprietary' instead. For more information, see https://spdx.org/licenses/.",
+            )
+            in emitter.interactions
+        )
+        # License should always remain unchanged
+        assert project.get("license") == lic
